requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
streamlit
python-dotenv
passlib
bcrypt
requests